requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
docker==6.1.0
PyYAML==6.0.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
numpy==1.24.3
